backend/live_agent.py
import os
from google import genai
from google.genai import types
from dotenv import load_dotenv

load_dotenv()

client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

async def process_audio_video_stream(data: dict):
    # In a real implementation, this would maintain a stateful connection
    # to the Gemini Live API using the gemini-2.5-flash-native-audio model.
    # For this prototype, we return a mock response.
    return {
        "type": "audio_response",
        "audio": "base64_encoded_pcm_audio_here"
    }
