requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
google-genai==0.3.0
playwright==1.40.0
